+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Gentics UI Core Changelog
 
-## 5.6.0 (in progress)
+## 5.6.0 (2017-11-22)
 
 ### Features